pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "valdi"
version = "0.1.0"
description = "A command-line utility for managing VALDI cloud resources"
authors = [{name = "VALDI", email="contact@valdi.ai"}]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.9"
keywords = ["valdi", "gpu", "cloud", "cli", "storage", "volumes"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology"
]
dependencies = [
    "certifi >= 2023.11.17",
    "charset-normalizer == 3.3.2",
    "idna >= 3.6",
    "PyJWT >= 2.8.0",
    "requests >= 2.31.0",
    "urllib3 >= 2.1.0"
]

[project.urls]
Homepage = "https://valdi.ai"
Documentation = "https://docs.valdi.ai"
Repository = "https://github.com/valdi-labs/valdi-mount-tool"

[project.scripts]
valdi = "valdi.entrypoint:main"